app probably done

+++ serverside/remove_tempfiles.py
@@ -1,4 +1,8 @@
 import config
 import os
+from database_handling.mongodb_wrapper import MongoAPI
 
-os.remove(config.basedir / "voter_dump.json")
+mongo = MongoAPI()
+if os.path.isfile(config.basedir / "voter_dump.json"):
+    os.remove(config.basedir / "voter_dump.json")
+mongo.clear_db()
